Refactor api/v1 to smaller go files
