+++ pkg/api/v1/assignments_param_binding.go
@@ -1,0 +1,37 @@
+package api
+
+import (
+	"github.com/labstack/echo/v4"
+	"github.com/volatiletech/sqlboiler/v4/queries/qm"
+
+	"go.infratographer.com/loadbalancerapi/internal/models"
+)
+
+func (r *Router) assignmentParamsBinding(c echo.Context) ([]qm.QueryMod, error) {
+	mods := []qm.QueryMod{}
+
+	tenantID, err := r.parseTenantID(c)
+	if err != nil {
+		return nil, err
+	}
+
+	mods = append(mods, models.AssignmentWhere.TenantID.EQ(tenantID))
+
+	queryParams := []string{"frontend_id", "load_balancer_id", "pool_id"}
+
+	qpb := echo.QueryParamsBinder(c)
+
+	for _, qp := range queryParams {
+		mods = queryParamsToQueryMods(qpb, qp, mods)
+
+		if len(c.QueryParam(qp)) > 0 {
+			r.logger.Debugw("query param", "query_param", qp, "param_vale", c.QueryParam(qp))
+		}
+	}
+
+	if err = qpb.BindError(); err != nil {
+		return nil, err
+	}
+
+	return mods, nil
+}
